Replace size check with static assert
The original check fails to compile on newer Ubuntus (20.04 and up)
in debug mode because it generates a signed conversion warning which
turns into an error thanks to `-Werror=sign-conversion` also being on.

We can get the same compile check with just a static assert, which
doesn't rely on questionable warning-inducing behaviour.